moved routes to separate object
